Set up Account and Resources pages.

+++ app/routes.py
@@ -1,10 +1,12 @@
-import os, secrets
+import os, random
 
 from flask import render_template, redirect, url_for, flash, request
 from app import app, db
-from app.forms import LoginForm, NewHelperForm, NewRecordForm
+from app.forms import (
+    LoginForm, NewHelperForm, NewHelperAdminForm, NewRecordForm, EditHelperForm, ResetPasswordForm, NewResourceForm, ContactAdminForm
+)
 from flask_login import current_user, login_user, logout_user, login_required
-from app.models import Resource, User, Circle, Record
+from app.models import Resource, User, Record, Circle
 from werkzeug.urls import url_parse
 
 
@@ -37,12 +39,19 @@ def logout():
 @login_required
 @app.route('/dashboard')
 def dashboard():
-    return render_template('dashboard.html')
+    helpers = len(User.query.all())
+    record_no = len(Record.query.all())
+    circles = len(Circle.query.all())
+    recent = Record.query.order_by(Record.id.desc()).limit(7)
+    pending = Record.query.filter_by(status="Processing").all()
+    return render_template('dashboard.html', helpers=helpers, record_no=record_no, circles=circles, recent=recent, pending=pending)
 
 
 @login_required
 @app.route('/helpers')
 def helpers():
+    if current_user.role == "Zonal Officer":
+        return redirect(url_for('login'))
     helpers = User.query.all()  
     return render_template('helpers.html', helpers=helpers)
 
@@ -50,38 +59,31 @@ def helpers():
 @login_required
 @app.route('/helper/new', methods=['GET', 'POST'])
 def new_helper():
-    form = NewHelperForm()
+    if current_user.role == "Admin":
+        form = NewHelperAdminForm()
+    elif current_user.role == "National Officer":
+        form = NewHelperForm
+    else:
+        return redirect(url_for('login'))
+    
     if request.method == "POST":
-        if form.validate_on_submit():
-            name = form.name.data
-            email = form.email.data
-            phone = form.phone.data
-            role = form.role.data
-            circle = form.circle.data
-            zone = form.zone.data
-                        
-            # The code below randomly combines 16 characters from the character_base variable to form a password.
-            character_base = '-./0123456789:;<=>?@ABCDEFGHIJKLMNOPQRSTUVWXYZ[\\]^_abcdefghijklmnopqrstuvwxyz%#$&*()+!}|{'
+        if form.validate_on_submit():                      
+            new_helper = User(
+                name=form.name.data, email=form.email.data, phone=form.phone.data, role=form.role.data, circle=form.circle.data, 
+                zone=form.zone.data
+            )            
+            # The code below generates a random password. 
+            character_base = 'abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!#$%&()*+/<=>@[\]{?}0123456789'
             password = ''
             for i in range(16):
-                password = password + secrets.choice(character_base)
+                password += random.choice(character_base)
+            print(password)
             
-            # Add new entries to the User and TemporaryPasswords tables.
-            new_helper = User(name=name, email=email, phone=phone, role=role, circle=circle, zone=zone)
             new_helper.set_password(password)
             db.session.add(new_helper)
             db.session.commit() 
-            flash(
-                f"New helper Mr. {name} was last added to the platform.",
-                category="success",
-            )
+            flash("New helper added successfully.", category="success")
             return redirect(url_for('helpers'))
-        else:
-            # flash(
-            #     f"Please fill all required form fields before submiting the form.",
-            #     category="danger",
-            # )
-            return render_template('new-helper.html', form=form)
     return render_template('new-helper.html', form=form)
     
 
@@ -89,6 +91,8 @@ def new_helper():
 @login_required
 @app.route('/helper/delete/<string:name>')
 def delete_helper(name):
+    if current_user.role == "Zonal Officer":
+        return redirect(url_for('login'))
     person = User.query.filter_by(name=name).first()
     if person:
         db.session.delete(person)
@@ -96,17 +100,57 @@ def delete_helper(name):
     return redirect(url_for('helpers'))
 
 
+@login_required
+@app.route('/helper/edit/<string:name>', methods=['GET', 'POST'])
+def edit_helper(name):
+    if current_user.role == "Zonal Officer":
+        return redirect(url_for('login'))
+    
+    user = User.query.filter_by(name=name).first()
+    form = EditHelperForm(obj=user)
+    if request.method == "POST":
+        if form.validate_on_submit():
+            user.phone = form.phone.data
+            user.email = form.email.data
+            user.circle = form.circle.data
+            user.zone = form.zone.data
+            user.role = form.role.data
+            db.session.commit()
+            flash(f'The record < {name} > has been updated', category='success')
+            return redirect(request.url)
+    return render_template("edit-helper.html", user=user, form=form)
+    
+
 
 @login_required
 @app.route('/resources')
 def resources():
-    resources = Resource.query.all()
-    return render_template('resources.html', resources=resources)
+    videos = Resource.query.filter_by(file_type="Video File").all()
+    pdfs = Resource.query.filter_by(file_type="PDF").all()
+    return render_template('resources.html', videos=videos, pdfs=pdfs)
+
+
+@login_required
+@app.route('/resources/new', methods=['GET', 'POST'])
+def new_resource():
+    form = NewResourceForm()
+    if request.method == 'POST':
+        if form.validate_on_submit():
+            new_item = Resource(
+                title=form.title.data, file_type=form.file_type.data, link=form.link.data
+            )
+            db.session.add(new_item)
+            db.session.commit()
+            flash(f'New resource added successfully', category='success')
+            return redirect(request.url)
+    return render_template('new-resource.html', form=form)
 
 
 @login_required
 @app.route('/records')
 def records():
+    if current_user.role == "Zonal Officer":
+        return redirect(url_for('login'))
     records = Record.query.all()
     return render_template('records.html', records=records)
 
@@ -114,30 +158,64 @@ def records():
 @login_required
 @app.route('/record/new', methods=['GET', 'POST'])
 def new_record():
+    if current_user.role == "Zonal Officer":
+        return redirect(url_for('login'))
     form = NewRecordForm()
     if request.method == "POST":
         if form.validate_on_submit():
-            name = form.name.data
-            circle = form.circle.data
-            zone = form.zone.data                  
-            dob = form.dob.data
-            dod = form.dod.data
-            yop = form.data.yop
-            cemetry = form.cemetry.data
-            
-            # Generate stencil and stencil path
-            
-            # Add new entries to the Record table.
-            new_record = User(name=name, circle=circle, zone=zone, dob=dob, dod=dod, yop=yop, cemetry=cemetry, stencil='', stencilPath='')
-            
+            new_record = Record(
+                name=form.name.data, circle=form.circle.data, zone=form.zone.data, dob=form.dob.data, dod=form.dod.data, yop=form.yop.data, cemetry=form.cemetry.data
+            )
             db.session.add(new_record)
+            
+            # Generate stencil and save to database
+            
             db.session.commit() 
             flash(
-                f"New tombstone record < {name} > was added to the platform.",
+                f"New tombstone record < {form.name.data} > was added to the platform.",
                 category="success",
             )
             return redirect(request.url)
-        else:
-            return render_template('new-record.html', form=form)
     return render_template('new-record.html', form=form)
-    
+
+
+@login_required
+@app.route('/reports')
+def reports():
+    return render_template('reports.html')
+
+
+@login_required
+@app.route('/account', methods=['GET', 'POST'])
+def account():
+    form = ResetPasswordForm()
+    user = User.query.filter_by(name=current_user.name).first()
+    if request.method == "POST":
+        if len(form.password2.data) < 16:
+            flash(f'Error! Your new password has less than 16 characters.', category='danger')
+            return redirect(request.url)
+        
+        if form.validate_on_submit():
+            user.set_password(form.password2.data)
+            db.session.commit()
+            flash(f'Your Password has been successfully reset', category='success')
+            return redirect(request.url)
+    return render_template('account.html', form=form)
+
+
+@login_required
+@app.route('/contact_admin')
+def contact_admin():
+    form = ContactAdminForm()
+    if request.method == 'POST':
+        if form.validate_on_submit():
+            # Send email
+            
+            flash(f'Your message was sent successfully!', category='success')
+            return redirect(request.url) 
+    return render_template('contact-admin.html', form=form)   
+
+
+# @app.errorhandler(404)
+# def not_found_error(error):
+#     return render_template("404.html"), 404
